requirements.txt
colorama==0.4.6
docopt==0.6.2
iniconfig==2.0.0
packaging==23.2
paho-mqtt==2.1.0
pluggy==1.3.0
pytest==7.4.3
pytest-watch==4.2.0
PyYAML==6.0.1
watchdog==3.0.0